requirements/requirements.txt
blinker
environs
Flask
Flask-Bcrypt
Flask-Bootstrap
Flask-DebugToolbar
Flask-Security
Flask-Migrate
Flask-SQLAlchemy
Flask-Testing
Flask-WTF
geoalchemy2
psycopg2-binary
sqlalchemy-utils
